backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.13.1
langchain-core>=0.3.70,<0.4
langgraph>=0.6.0
langchain-google-genai>=2.1.8,<3
google-ai-generativelanguage==0.6.18
python-dotenv==1.0.0
langgraph-prebuilt>=0.6.0
